requirements.txt
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.0.1
cfgv==3.4.0
click==8.2.1
distlib==0.3.9
fastapi==0.115.12
filelock==3.18.0
h11==0.16.0
identify==2.6.12
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
mysqlclient==2.2.5
nodeenv==1.9.1
passlib==1.7.4
platformdirs==4.3.8
pre_commit==4.2.0
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyJWT==2.9.0
python-dotenv==1.1.0
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.34.2
virtualenv==20.31.2
